+++ pkg/psr-queue/Producer.php
@@ -2,18 +2,9 @@
 
 namespace Enqueue\Psr;
 
-interface Producer
+/**
+ * @deprecated use PsrProducer
+ */
+interface Producer extends PsrProducer
 {
-    /**
-     * @param Destination $destination
-     * @param Message     $message
-     *
-     * @throws Exception                   - if the provider fails to send
-     *                                     the message due to some internal error
-     * @throws InvalidDestinationException - if a client uses
-     *                                     this method with an invalid destination
-     * @throws InvalidMessageException     - if an invalid message
-     *                                     is specified
-     */
-    public function send(Destination $destination, Message $message);
 }
