+++ pkg/psr-queue/PsrConnectionFactory.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Enqueue\Psr;
+
+interface PsrConnectionFactory
+{
+    /**
+     * @return Context
+     */
+    public function createContext();
+}
